# backend/internal/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: Home Finance API
  version: "1.4.0"
  description: |
    Internal API for a personal/family budgeting app. Plain net/http + chi
    router, hand-written JSON handlers. All money amounts are integer cents.
    Dates are YYYY-MM-DD strings; timestamps are RFC3339.

    Every error response uses the shape
    `{"error": {"code": "...", "message": "..."}}`. Messages are safe to show
    to a user; `db_error` carries a fixed generic message and never raw driver
    text.

    JSON request bodies are capped at 1 MiB by middleware; a larger body gets
    413 `body_too_large`. Multipart uploads bypass that cap and enforce their
    own, larger limit.

    Unmatched `/api` paths return 404 `not_found` and unsupported methods
    return 405 `method_not_allowed`, both as JSON rather than the SPA.

    Operations marked "Admin only" also require the caller's role to be
    `admin`; other roles get 403 `forbidden`.
servers:
  - url: /
...
security:
  - cookieAuth: []
    csrfHeader: []
paths:
...
  /api/years:
    get:
      operationId: listYears
      tags: [Years]
      summary: Lists all known years, ascending.
      responses:
        "200":
          description: Years.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createYear
      tags: [Years]
      summary: Registers a year (2000-2100), idempotent. Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                year:
                  type: integer
      responses:
        "201":
          description: Year registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  year:
                    type: integer
        "400":
          description: Invalid body or year out of range.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/years/{year}/lock:
    post:
      operationId: lockYear
      tags: [Years]
      summary: Locks a year after a fresh Authentik re-authentication, once all its periods are closed. Admin only.
      parameters:
        - $ref: "#/components/parameters/YearParam"
      responses:
        "204":
          description: Year locked.
        "400":
          description: Invalid year.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: No fresh re-authentication on file; client should trigger the /auth/reauth popup and retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Year already locked, or open periods still exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/years/{year}/unlock:
    post:
      operationId: unlockYear
      tags: [Years]
      summary: Unlocks a year after a fresh Authentik re-authentication. Admin only.
      parameters:
        - $ref: "#/components/parameters/YearParam"
      responses:
        "204":
          description: Year unlocked.
        "400":
          description: Invalid year.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: No fresh re-authentication on file; client should trigger the /auth/reauth popup and retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          description: Server/DB error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set after successful login (OIDC or dev fake auth).
    csrfHeader:
      type: apiKey
      in: header
      name: X-Requested-With
      description: Required (value "XMLHttpRequest") on all mutating /api requests alongside the session cookie.
  parameters:
...
    YearParam:
      name: year
      in: path
      required: true
      schema:
        type: integer
  responses:
...
    Forbidden:
      description: Admin role required, or the X-Requested-With CSRF header is missing (code `forbidden` / `csrf`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BodyTooLarge:
      description: JSON body exceeded the 1 MiB limit (code `body_too_large`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...